removed duplicate line in chart-remove-allowZoom,  due to running 'create' twice renamed rule
