+++ packages/eslint-plugin-pf-codemods/lib/rules/chart-remove-allowZoom.js
